Typecast offset and line height to int.

+++ src/draw.c
@@ -6,7 +6,7 @@
 /*   By: msousa <mlrcbsousa@gmail.com>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 12:06:29 by msousa            #+#    #+#             */
-/*   Updated: 2022/05/09 01:23:14 by msousa           ###   ########.fr       */
+/*   Updated: 2022/05/09 16:05:04 by msousa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,32 +27,12 @@ int	create_trgb(int t, int r, int g, int b)
 
 void	draw(t_app *self)
 {
-	// int	i;
-	// int	j;
-	// int	color;
-
-	// i = 0;
-	// while (i < WIDTH)
-	// {
-	// 	j = 0;
-	// 	while (j < HEIGHT)
-	// 	{
-	// 		color = create_trgb(0, 0, 0, 0);
-	// 		my_mlx_pixel_put(self->img, i, j, color);
-	// 		j++;
-	// 	}
-	// 	i++;
-	// }
 	draw_rays(self);
 	mlx_put_image_to_window(self->mlx, self->mlx_window, self->img->img, 0, 0);
 }
 
-
-void	draw_line(t_app *self, int ray, float line_height, float line_offset)
+void	draw_line(t_app *self, int ray, int line_height, int line_offset)
 {
-	// printf("drawing line\n");
-	// printf("ray: %d, line_height: %f, line_offset: %f\n", ray, line_height, line_offset);
-
 	int	j;
 	int	color;
 
@@ -65,7 +45,7 @@ void	draw_line(t_app *self, int ray, float line_height, float line_offset)
 	}
 	while (j < line_offset + line_height)
 	{
-		// TODO: decide which wall to point
+		// TODO: decide which wall to paint
 		my_mlx_pixel_put(self->img, ray, j, g_wall_color);
 		j++;
 	}
